Improving genre search started from the manga page of a popular manga
Tapping on a genre while viewing a popular manga of a HttpSource/ParsedHttpSource now brings up better results,
if the extension developer implemented a genre filter into the extension filter list.

Works with genre filters implemented as groups of checkbox filters/tri-state filters, as well as with simple drop down filters.

Extension developers don't have to do any additional work beyond implementing the genre filter properly.
The genre in the filter must have the same name as the genre that show up below the manga description (upper/lower case ignored).

The new genre filter method has a built-in fallback to the old way tapping on a genre was handled:
If a genre name can't be found in the extension filters, it'll simply search for the genre as if it is a manga title.